+++ footer.php
@@ -14,8 +14,10 @@
 		</div>
 		<!-- /Wrapper -->
 		
-		<!-- jQuery Fallback -->
-		<script>window.jQuery || document.write('<script src="http://cdnjs.cloudflare.com/ajax/libs/jquery/1.9.1/jquery.min.js"><\/script>');</script>
+		<!-- jQuery CDN and Custom Scripts -->
+		<script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+		<script>window.jQuery || document.write('<script src="http://cdnjs.cloudflare.com/ajax/libs/jquery/1.9.1/jquery.min.js"><\/script>')</script>
+		<?php wp_footer(); ?>
 		
 		<!-- Analytics -->
 		<script>
@@ -24,8 +26,6 @@
 			g.src=('https:'==location.protocol?'//ssl':'//www')+'.google-analytics.com/ga.js';
 			s.parentNode.insertBefore(g,s)})(document,'script');
 		</script>
-		
-		<?php wp_footer(); ?>
 	
 	</body>
 </html>